Conditional service reconfiguration

+++ kickstart/scripts/hotspot-deploy.sh
@@ -27,21 +27,27 @@ deploy_hotspot_ubuntu() {
 
     systemctl restart systemd-networkd
 
-    expand etc/hostapd.conf "/etc/hostapd/hostapd.conf"
-    expand etc/dnsmasq-posm.conf "/etc/dnsmasq.d/50-posm.conf"
+    # configure hostapd / dnsmasq if appropriate
+    if [ -f /etc/hostapd/hostapd.conf ]; then
+      expand etc/hostapd.conf "/etc/hostapd/hostapd.conf"
+      expand etc/dnsmasq-posm.conf "/etc/dnsmasq.d/50-posm.conf"
 
-    ln -sf /run/systemd/resolve/resolv.conf /etc/resolv.conf
+      ln -sf /run/systemd/resolve/resolv.conf /etc/resolv.conf
 
-    grep -qe "^DAEMON_CONF" /etc/default/hostapd || echo DAEMON_CONF=\"/etc/hostapd/hostapd.conf\" >> /etc/default/hostapd
+      grep -qe "^DAEMON_CONF" /etc/default/hostapd || echo DAEMON_CONF=\"/etc/hostapd/hostapd.conf\" >> /etc/default/hostapd
 
-    systemctl unmask hostapd.service
-    systemctl enable hostapd.service
-    systemctl start hostapd.service
-    systemctl restart dnsmasq.service
+      systemctl unmask hostapd.service
+      systemctl enable hostapd.service
+      systemctl start hostapd.service
+      systemctl restart dnsmasq.service
+    fi
 
-    # rewrite Samba config to include active interfaces
-    expand etc/smb.conf /etc/samba/smb.conf
-    service smbd restart
+    # reconfigure samba if installed
+    if [ -f /etc/samba/smb.conf ]; then
+      # rewrite Samba config to include active interfaces
+      expand etc/smb.conf /etc/samba/smb.conf
+      service smbd restart
+    fi
   else
     # configure the VM's WAN interface
     expand etc/systemd/network/wan.network.hbs /etc/systemd/network/10-wan.network
